replace http with dio
